Test refactors, fix off-by-ones
Valid positions in a piece are as follows

  "W o r l d"
  0 1 2 3 4

e.g. the position is just before the character.  This means that
the end of that piece is at character 4.  If you traverse to the
right, you don't go to pos 5 (to the right of 'd) but rather to pos
0 of the *next* piece.

(This means that if you :bounce on the :right, it is a little odd as it
currently stands, because you have to stay at pos 4, which is to the
left of the final character.  We may need some special handling here,
like an End-of-File marker?  But this is better than having a special
case for positioning.)

Note that we also go back to the 4th pos (not the inexistent 5th pos)
when traversing *back* to that piece.

Also refactored test-dactyl-at into a macro, to preserve caller-local
line number on failure.